Rename interface_types back to types
