+++ packages/teams_memory/teams_memory/interfaces/types.py
